requirements.txt
python-dotenv==1.0.1
pandas==2.0.3
numpy==1.21.0
requests==2.32.3
PyJWT==2.9.0
pandas-gbq==0.23.1
tqdm==4.66.5